+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Changelog
   - `--apply` option has been removed as it is the default behaviour.
   - isort now does nothing, beyond giving instructions and exiting status code 0, when ran with no arguments.
     - a new `--interactive` flag has been added to enable the old style behaviour.
+  - isort now works on contiguous sections of imports, instead of one whole file at a time.
 
 Internal:
   - isort now utilizes mypy and typing to filter out typing related issues before deployment.